User login/Registration and JWT auth request with relationship between users and contacts

+++ controllers/contactController.js
@@ -2,16 +2,16 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id : req.user.id});
     res.status(200).json({contacts});
     // res.status(200).json({message:"get all contacts"});
 });
 
 //@desc Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     console.log("the request body is :",req.body);
     const {name, email, phone} = req.body;
@@ -23,6 +23,7 @@ const createContact = asyncHandler(async (req, res) => {
         name,
         email,
         phone,
+        user_id : req.user.id
     });
     res.status(201).json(contact);
     // res.status(201).json({message:"create contacts"});
@@ -30,7 +31,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -44,12 +45,17 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
     }
 
     const updateContact = await Contact.findByIdAndUpdate(
@@ -63,13 +69,19 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("contact not found");
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user contacts");
+    }
+    
     await Contact.findByIdAndRemove(req.params.id);
     res.status(200).json(contact);
     // res.status(200).json({message:`delete contact for ${req.params.id}`});
